# records/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "School Academic Records Facade",
    "description": "Read-only academic records for the school assistant agent, plus the guardian authorisation and audit trail that govern them.\n\n**Two independent credentials are required for every parent-facing read.** The `X-API-Key` header proves which *system* is calling. An `Authorization: Bearer` identity token, signed by the identity service, proves *which parent* it asks on behalf of — and its `guardian_id` claim must match the `guardian_id` in the path. Neither alone grants anything: a leaked API key cannot choose a guardian, and a parent's token cannot be presented without a valid system key. The permitted student set is then resolved server-side from the link table on every request, and is never supplied by the caller.\n\n**On failure, do not improvise.** A 503 with `code: lms_unavailable` means the system of record could not be reached. The correct response to a parent is that records are temporarily unavailable, never a remembered or inferred figure.",
    "version": "0.1.0"
  },
  "paths": {
    "/v1/terms": {
      "get": {
        "tags": [
          "records"
        ],
        "summary": "List Terms",
        "description": "Terms the school has configured. Carries no student data, so needs no subject.",
        "operationId": "list_terms_v1_terms_get",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TermOut"
                  },
                  "title": "Response List Terms V1 Terms Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/guardians/{guardian_id}/students": {
      "get": {
        "tags": [
          "records"
        ],
        "summary": "List Students",
        "description": "Children this guardian may ask about.\n\nThe agent's first call in any conversation. An unknown guardian and a guardian\nwith no visible children both return an empty list — see `permitted_students`.",
        "operationId": "list_students_v1_guardians__guardian_id__students_get",
        "parameters": [
          {
            "name": "guardian_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Guardian Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentListOut"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key, or missing/invalid identity token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "Identity token does not authorise the guardian named in the path.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "No such student record for this guardian.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "503": {
            "description": "System of record unreachable. Do not answer from memory.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/guardians/{guardian_id}/students/{student_id}/grades": {
      "get": {
        "tags": [
          "records"
        ],
        "summary": "Get Grades",
        "description": "Every subject's rollup for one student in one term.",
        "operationId": "get_grades_v1_guardians__guardian_id__students__student_id__grades_get",
        "parameters": [
          {
            "name": "guardian_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Guardian Id"
            }
          },
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Student Id"
            }
          },
          {
            "name": "term",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Term code; defaults to the current term.",
              "title": "Term"
            },
            "description": "Term code; defaults to the current term."
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentGradesOut"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key, or missing/invalid identity token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "Identity token does not authorise the guardian named in the path.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "No such student record for this guardian.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "503": {
            "description": "System of record unreachable. Do not answer from memory.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/admin/guardians": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Create Guardian",
        "operationId": "create_guardian_v1_admin_guardians_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuardianIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Guardian V1 Admin Guardians Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorOut": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "not_authorized | unknown_student | unknown_term | records_restricted | lms_unavailable | not_found"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "ErrorOut",
        "description": "Machine-readable failure.\n\n`code` exists so the agent branches on a value rather than on prose. The one that\nmatters most is `lms_unavailable`: it must produce \"I cannot reach the school\nrecords right now\", never an answer the model invented to be helpful."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}